Edit functinality solve reset problem

+++ 1st_TodoList_Redux/src/App.jsx
@@ -11,7 +11,12 @@ function App() {
   const [task, setTask] = useState("");
 
   const handleDelete = (id) => {
-    dispatch(deleteTodo(id));
+    const confirm = window.confirm(
+      "Are you sure you want to delete this task?"
+    );
+    if (confirm) {
+      dispatch(deleteTodo(id));
+    }
   };
 
   const handleUpdate = (id, title) => {
@@ -21,7 +26,7 @@ function App() {
   return (
     <div className="w-10/12 md:w-1/2 lg:w-1/3 h-auto p-10 mt-10 text-stone-900 text-center mx-auto bg-gray-50 rounded-lg shadow-lg">
       <h1 className="text-center text-orange-500 text-xl">Todo List</h1>
-      <TodoForm task={task} />
+      <TodoForm task={task} setTask={setTask} />
       <div className="mt-5">
         {todos.map((todo) => (
           <div
